+++ src/oci/integration/models/update_integration_instance_details.py
@@ -37,6 +37,10 @@ class UpdateIntegrationInstanceDetails(object):
     #: This constant has a value of "ENTERPRISEX"
     INTEGRATION_INSTANCE_TYPE_ENTERPRISEX = "ENTERPRISEX"
 
+    #: A constant which can be used with the integration_instance_type property of a UpdateIntegrationInstanceDetails.
+    #: This constant has a value of "HEALTHCARE"
+    INTEGRATION_INSTANCE_TYPE_HEALTHCARE = "HEALTHCARE"
+
     def __init__(self, **kwargs):
         """
         Initializes a new UpdateIntegrationInstanceDetails object with values from keyword arguments.
@@ -48,7 +52,7 @@ class UpdateIntegrationInstanceDetails(object):
 
         :param integration_instance_type:
             The value to assign to the integration_instance_type property of this UpdateIntegrationInstanceDetails.
-            Allowed values for this property are: "STANDARD", "ENTERPRISE", "STANDARDX", "ENTERPRISEX"
+            Allowed values for this property are: "STANDARD", "ENTERPRISE", "STANDARDX", "ENTERPRISEX", "HEALTHCARE"
         :type integration_instance_type: str
 
         :param freeform_tags:
@@ -153,7 +157,7 @@ class UpdateIntegrationInstanceDetails(object):
         Oracle Integration Generation 2 uses ENTERPRISE and STANDARD,
         Oracle Integration 3 uses ENTERPRISEX and STANDARDX
 
-        Allowed values for this property are: "STANDARD", "ENTERPRISE", "STANDARDX", "ENTERPRISEX"
+        Allowed values for this property are: "STANDARD", "ENTERPRISE", "STANDARDX", "ENTERPRISEX", "HEALTHCARE"
 
 
         :return: The integration_instance_type of this UpdateIntegrationInstanceDetails.
@@ -173,7 +177,7 @@ class UpdateIntegrationInstanceDetails(object):
         :param integration_instance_type: The integration_instance_type of this UpdateIntegrationInstanceDetails.
         :type: str
         """
-        allowed_values = ["STANDARD", "ENTERPRISE", "STANDARDX", "ENTERPRISEX"]
+        allowed_values = ["STANDARD", "ENTERPRISE", "STANDARDX", "ENTERPRISEX", "HEALTHCARE"]
         if not value_allowed_none_or_none_sentinel(integration_instance_type, allowed_values):
             raise ValueError(
                 f"Invalid value for `integration_instance_type`, must be None or one of {allowed_values}"