+++ src/oci/network_load_balancer/models/backend.py
@@ -15,7 +15,7 @@ class Backend(object):
     The configuration of a backend server that is a member of a network load balancer backend set.
     For more information, see `Managing Backend Servers`__.
 
-    __ https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/BackendServers/backend-server-management.htm
+    __ https://docs.cloud.oracle.com/Content/Balance/Tasks/managingbackendservers.htm
     """
 
     def __init__(self, **kwargs):
@@ -207,7 +207,7 @@ class Backend(object):
 
         Example: `3`
 
-        __ https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/introducton.htm#Policies
+        __ https://docs.cloud.oracle.com/Content/Balance/Reference/lbpolicies.htm
 
 
         :return: The weight of this Backend.
@@ -227,7 +227,7 @@ class Backend(object):
 
         Example: `3`
 
-        __ https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/introducton.htm#Policies
+        __ https://docs.cloud.oracle.com/Content/Balance/Reference/lbpolicies.htm
 
 
         :param weight: The weight of this Backend.