+++ src/oci/network_load_balancer/models/listener.py
@@ -13,10 +13,10 @@ from oci.decorators import init_model_state_from_kwargs
 class Listener(object):
     """
     The congfiguration of the listener.
-    For more information about listener configuration, see
+    For more information about backend set configuration, see
     `Managing Load Balancer Listeners`__.
 
-    __ https://docs.cloud.oracle.com/Content/NetworkLoadBalancer/Listeners/listener-management.htm
+    __ https://docs.cloud.oracle.com/Content/Balance/Tasks/managinglisteners.htm
     """
 
     #: A constant which can be used with the protocol property of a Listener.
@@ -76,6 +76,14 @@ class Listener(object):
             The value to assign to the is_ppv2_enabled property of this Listener.
         :type is_ppv2_enabled: bool
 
+        :param tcp_idle_timeout:
+            The value to assign to the tcp_idle_timeout property of this Listener.
+        :type tcp_idle_timeout: int
+
+        :param udp_idle_timeout:
+            The value to assign to the udp_idle_timeout property of this Listener.
+        :type udp_idle_timeout: int
+
         """
         self.swagger_types = {
             'name': 'str',
@@ -83,7 +91,9 @@ class Listener(object):
             'port': 'int',
             'protocol': 'str',
             'ip_version': 'str',
-            'is_ppv2_enabled': 'bool'
+            'is_ppv2_enabled': 'bool',
+            'tcp_idle_timeout': 'int',
+            'udp_idle_timeout': 'int'
         }
 
         self.attribute_map = {
@@ -92,7 +102,9 @@ class Listener(object):
             'port': 'port',
             'protocol': 'protocol',
             'ip_version': 'ipVersion',
-            'is_ppv2_enabled': 'isPpv2Enabled'
+            'is_ppv2_enabled': 'isPpv2Enabled',
+            'tcp_idle_timeout': 'tcpIdleTimeout',
+            'udp_idle_timeout': 'udpIdleTimeout'
         }
 
         self._name = None
@@ -101,6 +113,8 @@ class Listener(object):
         self._protocol = None
         self._ip_version = None
         self._is_ppv2_enabled = None
+        self._tcp_idle_timeout = None
+        self._udp_idle_timeout = None
 
     @property
     def name(self):
@@ -279,6 +293,58 @@ class Listener(object):
         :type: bool
         """
         self._is_ppv2_enabled = is_ppv2_enabled
+
+    @property
+    def tcp_idle_timeout(self):
+        """
+        Gets the tcp_idle_timeout of this Listener.
+        The duration for TCP idle timeout in seconds.
+        Example: `300`
+
+
+        :return: The tcp_idle_timeout of this Listener.
+        :rtype: int
+        """
+        return self._tcp_idle_timeout
+
+    @tcp_idle_timeout.setter
+    def tcp_idle_timeout(self, tcp_idle_timeout):
+        """
+        Sets the tcp_idle_timeout of this Listener.
+        The duration for TCP idle timeout in seconds.
+        Example: `300`
+
+
+        :param tcp_idle_timeout: The tcp_idle_timeout of this Listener.
+        :type: int
+        """
+        self._tcp_idle_timeout = tcp_idle_timeout
+
+    @property
+    def udp_idle_timeout(self):
+        """
+        Gets the udp_idle_timeout of this Listener.
+        The duration for UDP idle timeout in seconds.
+        Example: `120`
+
+
+        :return: The udp_idle_timeout of this Listener.
+        :rtype: int
+        """
+        return self._udp_idle_timeout
+
+    @udp_idle_timeout.setter
+    def udp_idle_timeout(self, udp_idle_timeout):
+        """
+        Sets the udp_idle_timeout of this Listener.
+        The duration for UDP idle timeout in seconds.
+        Example: `120`
+
+
+        :param udp_idle_timeout: The udp_idle_timeout of this Listener.
+        :type: int
+        """
+        self._udp_idle_timeout = udp_idle_timeout
 
     def __repr__(self):
         return formatted_flat_dict(self)